+++ .eslintrc.js
@@ -13,4 +13,20 @@ module.exports = {
 		'plugin:@typescript-eslint/eslint-recommended',
 		'plugin:@typescript-eslint/recommended',
 	],
+	overrides: [
+		{
+			files: ['packages/{pac-proxy,proxy,socks-proxy}-agent/**/*'],
+			rules: {
+				'no-restricted-globals': [
+					'error',
+					// https://github.com/TooTallNate/proxy-agents/pull/242
+					{
+						name: 'URL',
+						message:
+							'Use `URL` from the Node.js "url" module instead.',
+					},
+				],
+			},
+		},
+	],
 };