+++ packages/proxy-agent/src/index.ts
@@ -1,5 +1,6 @@
 import * as http from 'http';
 import * as https from 'https';
+import { URL } from 'url';
 import LRUCache from 'lru-cache';
 import { Agent, AgentConnectOpts } from 'agent-base';
 import createDebug from 'debug';
@@ -36,11 +37,11 @@ export const proxies: {
 	socks4a: [SocksProxyAgent, SocksProxyAgent],
 	socks5: [SocksProxyAgent, SocksProxyAgent],
 	socks5h: [SocksProxyAgent, SocksProxyAgent],
-	'pac-data': [PacProxyAgent, PacProxyAgent],
-	'pac-file': [PacProxyAgent, PacProxyAgent],
-	'pac-ftp': [PacProxyAgent, PacProxyAgent],
-	'pac-http': [PacProxyAgent, PacProxyAgent],
-	'pac-https': [PacProxyAgent, PacProxyAgent],
+	'pac+data': [PacProxyAgent, PacProxyAgent],
+	'pac+file': [PacProxyAgent, PacProxyAgent],
+	'pac+ftp': [PacProxyAgent, PacProxyAgent],
+	'pac+http': [PacProxyAgent, PacProxyAgent],
+	'pac+https': [PacProxyAgent, PacProxyAgent],
 };
 
 function isValidProtocol(v: string): v is ValidProtocol {